Fixed compilation warnings under GCC

+++ 06-texture/06-texture.cpp
@@ -111,7 +111,10 @@ public:
         });
         // Setup texture data description
         const VkFormat format = utilities::getBlockCompressedFormat(ctx);
-        const VkExtent2D extent = {ctx.image_width(0, 0), ctx.image_height(0, 0)};
+        const VkExtent2D extent = {
+            (uint32_t)ctx.image_width(0, 0),
+            (uint32_t)ctx.image_height(0, 0)
+        };
         magma::Image::MipmapLayout mipOffsets(1, 0);
         for (int level = 1; level < ctx.num_mipmaps(0); ++level)
         {   // Compute relative offset
